init: be a bit more careful about what is added to local-package-dirs
The MELPA recipes dirs contains a .dir-locals.el plus a number of files that contain sexps but are not elisp
code.  Adding it to the list meant that eg (load "async") would try to load the async recipe as an elisp file
which gave me a bad time.  Need to sort this whole local packages thing out properly when I have time.